apps/roblox-animation/pyproject.toml
[project]
name = "roblox-animation"
version = "0.1.0"
description = "CLI tools for Roblox NPC animation with Gemini vision analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "pillow>=10.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
analyze-image = "gemini_analyzer.single_image:main"
analyze-frames = "gemini_analyzer.frame_sequence:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gemini_analyzer", "src/blender_scripts"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
